added note deletion confirm

+++ frontend/src/NotesList/Note/Buttons/DeleteNoteButton.jsx
@@ -2,6 +2,9 @@ import { StyledButton } from "./Buttons";
 
 function deleteNote(event, props) {
     event.preventDefault();
+    if (window.confirm("Sure you want to delete this note?") === false) {
+        return;
+    }
     const url = props.url + "/delete/" + props.note.id;
     const options = {
         method: "DELETE",
